Mopdule rename & code cleaning
