+++ RNBridgeApp/ios/RNBridgeApp-Bridging-Header.h
@@ -2,10 +2,5 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#ifndef RNBridgeApp-Bridging-Header
-#define RNBridgeApp-Bridging-Header
-
-
-#endif
 
 #import "React/RCTBridgeModule.h"
